+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.4.1] - 2025-09-20
 
 ### 🚀 Features
 
+- Push on install and uninstall
 - Improved installation via metadata
 - Show count